+++ source/renderer/app/api/staking/types.js
@@ -1,22 +1,27 @@
 // @flow
 import BigNumber from 'bignumber.js';
+import StakePool from '../../domains/StakePool';
 
-export type StakePool = {
+export type AdaApiStakePool = {
   id: string,
-  controlledStake: number,
-  description: string,
-  slug: string,
-  name: string,
-  performance: number,
-  profitMargin: number,
-  ranking: number,
-  retiring?: Date,
-  created_at: Date,
-  isCharity: boolean,
-  url: string,
+  metrics: {
+    controlled_stake: {
+      quantity: number,
+      unit: 'lovelace',
+    },
+    produced_blocks: {
+      quantity: number,
+      unit: 'block',
+    },
+  },
+  apparent_performance: number,
+  metadata: {
+    ticker: string,
+    homepage: string,
+    pledge_address: string,
+  },
 };
-
-export type StakePoolsListType = Array<StakePool>;
+export type AdaApiStakePools = Array<AdaApiStakePool>;
 
 export type Reward = {
   date: string,