+++ source/renderer/app/components/wallet/transactions/WalletTransactionsList.js
@@ -12,7 +12,6 @@ import styles from './WalletTransactionsList.scss';
 import Transaction from './Transaction';
 import { WalletTransaction } from '../../../domains/WalletTransaction';
 import LoadingSpinner from '../../widgets/LoadingSpinner';
-import { DEVELOPMENT } from '../../../../../common/types/environment.types';
 import { VirtualTransactionList } from './render-strategies/VirtualTransactionList';
 import { TransactionInfo, TransactionsGroup } from './types';
 import type { Row } from './types';
@@ -50,15 +49,13 @@ type Props = {
   isLoadingTransactions: boolean,
   isRestoreActive: boolean,
   isRenderingAsVirtualList: boolean,
-  network: string,
-  rawNetwork: string,
   onShowMoreTransactions?: Function,
-  onOpenExternalLink?: Function,
+  onOpenExternalLink: Function,
+  getUrlByType: Function,
   showMoreTransactionsButton?: boolean,
   transactions: Array<WalletTransaction>,
   walletId: string,
   isDeletingTransaction: boolean,
-  currentLocale: string,
   currentDateFormat: string,
   currentTimeFormat: string,
 };
@@ -73,8 +70,6 @@ export default class WalletTransactionsList extends Component<Props> {
 
   static defaultProps = {
     isRenderingAsVirtualList: false,
-    network: DEVELOPMENT,
-    rawNetwork: DEVELOPMENT,
     showMoreTransactionsButton: false,
     onShowMoreTransactions: () => {},
     onOpenExternalLink: () => {},
@@ -181,12 +176,10 @@ export default class WalletTransactionsList extends Component<Props> {
       deletePendingTransaction,
       formattedWalletAmount,
       isRestoreActive,
-      network,
-      rawNetwork,
       onOpenExternalLink,
+      getUrlByType,
       walletId,
       isDeletingTransaction,
-      currentLocale,
       currentTimeFormat,
     } = this.props;
     const { isFirstInGroup, isLastInGroup, tx } = data;
@@ -204,14 +197,12 @@ export default class WalletTransactionsList extends Component<Props> {
           isExpanded={this.isTxExpanded(tx)}
           isLastInList={isLastInGroup}
           isRestoreActive={isRestoreActive}
-          network={network}
-          rawNetwork={rawNetwork}
           onDetailsToggled={() => this.toggleTransactionExpandedState(tx)}
           onOpenExternalLink={onOpenExternalLink}
+          getUrlByType={getUrlByType}
           state={tx.state}
           walletId={walletId}
           isDeletingTransaction={isDeletingTransaction}
-          currentLocale={currentLocale}
           currentTimeFormat={currentTimeFormat}
         />
       </div>